:sparkles: cleanup stale label in v2 scaffolding
