+++ pkg/scaffold/project.go
@@ -115,7 +115,7 @@ func (p *V1Project) Scaffold() error {
 		&scaffoldv1.KustomizeImagePatch{},
 		&metricsauthv1.KustomizePrometheusMetricsPatch{},
 		&metricsauthv1.KustomizeAuthProxyPatch{},
-		&project.AuthProxyService{},
+		&scaffoldv1.AuthProxyService{},
 		&project.AuthProxyRole{},
 		&project.AuthProxyRoleBinding{},
 		&manager.Config{Image: imgName},
@@ -184,7 +184,7 @@ func (p *V2Project) Scaffold() error {
 		&scaffoldv2.KustomizeImagePatch{},
 		&metricsauthv2.KustomizePrometheusMetricsPatch{},
 		&metricsauthv2.KustomizeAuthProxyPatch{},
-		&project.AuthProxyService{},
+		&scaffoldv2.AuthProxyService{},
 		&project.AuthProxyRole{},
 		&project.AuthProxyRoleBinding{},
 		&managerv2.Config{Image: imgName},
