+++ pkg/scaffold/v1/authproxyservice.go
@@ -14,52 +14,48 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resource
+package v1
 
 import (
-	"fmt"
 	"path/filepath"
 
 	"sigs.k8s.io/kubebuilder/pkg/scaffold/input"
 )
 
-var _ input.File = &RoleBinding{}
+var _ input.File = &AuthProxyService{}
 
-// RoleBinding scaffolds the config/manager/group_rolebinding_rbac.yaml file
-type RoleBinding struct {
+// AuthProxyService scaffolds the config/rbac/auth_proxy_service.yaml file
+type AuthProxyService struct {
 	input.Input
-
-	// Resource is a resource in the API group
-	Resource *Resource
 }
 
 // GetInput implements input.File
-func (r *RoleBinding) GetInput() (input.Input, error) {
+func (r *AuthProxyService) GetInput() (input.Input, error) {
 	if r.Path == "" {
-		r.Path = filepath.Join("config", "manager", fmt.Sprintf(
-			"%s_rolebinding_rbac.yaml", r.Resource.Group))
+		r.Path = filepath.Join("config", "rbac", "auth_proxy_service.yaml")
 	}
-	r.TemplateBody = roleBindingTemplate
+	r.TemplateBody = AuthProxyServiceTemplate
 	return r.Input, nil
 }
 
-// Validate validates the values
-func (r *RoleBinding) Validate() error {
-	return r.Resource.Validate()
-}
-
-var roleBindingTemplate = `apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
+var AuthProxyServiceTemplate = `apiVersion: v1
+kind: Service
 metadata:
+  annotations:
+    prometheus.io/port: "8443"
+    prometheus.io/scheme: https
+    prometheus.io/scrape: "true"
   labels:
+    control-plane: controller-manager
     controller-tools.k8s.io: "1.0"
-  name: {{ .Resource.Group }}-rolebinding
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: {{ .Resource.Group }}-role
-subjects:
-- kind: ServiceAccount
-  name: default
+  name: controller-manager-metrics-service
   namespace: system
+spec:
+  ports:
+  - name: https
+    port: 8443
+    targetPort: https
+  selector:
+    control-plane: controller-manager
+    controller-tools.k8s.io: "1.0"
 `
