+++ pkg/scaffold/v2/crd/enablecainjection_patch.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v2
+package crd
 
 import (
 	"fmt"
