+++ pkg/scaffold/v2/webhook/service.go
@@ -51,5 +51,4 @@ spec:
       targetPort: 443
   selector:
     control-plane: controller-manager
-    controller-tools.k8s.io: "1.0"
 `
