+++ testdata/project-v2/config/webhook/service.yaml
@@ -10,4 +10,3 @@ spec:
       targetPort: 443
   selector:
     control-plane: controller-manager
-    controller-tools.k8s.io: "1.0"
